source refactoring

+++ routes/register-teacher-profile.js
@@ -5,12 +5,6 @@ var UserMultiAttrDAO = require('../db/user-multi-attr-dao')
 class RegisterTeacherProfileRouter extends BeauEduRouter {
   async processRequest () {
 
-    var result = false;
-    if (true !== this.req.session.logined) {
-      this.res.render('login.ejs')
-      return result
-    }
-
     var data = {}
     data.session = this.req.session
     data.user_id = this.req.session.user_id
@@ -31,17 +25,19 @@ class RegisterTeacherProfileRouter extends BeauEduRouter {
         returnVal = await teacherProfileDAO.updateTeacherProfile(1)
       }
 
+      if (0 == returnVal.rowCount) throw 'Register teacher profile failed.'
+
       var attrVal
       var teacherSpeciality = this.inputParam.teacher_specialities
 
       returnVal = await userMultiAttrDAO.deleteUserMultiAttr(data.user_id, 'teacher_speciality')
       var key
       for (key in teacherSpeciality) {
-
         attrVal = teacherSpeciality[key]
-
         returnVal = await userMultiAttrDAO.insertUserMultiAttr(data.user_id, 'teacher_speciality', attrVal)
       }
+
+      if (0 == returnVal.rowCount) throw 'Register teacher profile failed.'
 
       data.result = 'success'
       await this.commit()
@@ -49,7 +45,7 @@ class RegisterTeacherProfileRouter extends BeauEduRouter {
       this.req.session.last_name = this.inputParam.last_name
       this.req.session.email = this.inputParam.email
       this.json(data)
-      result = true
+      return true
 
     } catch (e) {
       await this.rollback()
@@ -57,11 +53,10 @@ class RegisterTeacherProfileRouter extends BeauEduRouter {
       this.json(data);
       this.error(e.stack || e)
       this.res.render('error.ejs')
-      result = false
+      return false
+
     } finally {
       if (null !== this.conn) await this.conn.release()
-
-      return result
     }
   }  
 }
